Add required changes
