+++ package.json
@@ -5,7 +5,8 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "build": "webpack --mode=production",
+    "dev": "webpack --mode development",
+    "build": "webpack --mode production",
     "webpack": "npx webpack --mode development --watch"
   },
   "keywords": [],
